[arcane] Déplace la définition de 'IMeshMaterialMng' dans 'arcane/core/materials'
