+++ arcane/ceapart/src/arcane/materials/srcs.cmake
@@ -129,7 +129,6 @@ set(ARCANE_SOURCES
   IMeshEnvironment.h
   IMeshMaterialVariableSynchronizer.h
   IMeshMaterialMng.h
-  IMeshMaterialMng.cc
   IMeshMaterialVariable.h
   IMeshMaterialModifierImpl.h
   IMeshMaterialVariableComputeFunction.h
